requirements.txt
rpi.gpio
wiringpi
